feat: class 종료 제외
