+++ src/components/class/VideoHeader.jsx
@@ -3,20 +3,20 @@ import { useState, useEffect, useCallback } from "react";
 import styled from "styled-components";
 
 import { apis } from "../../shared/axios";
-import { ST2 } from "../../../styles/Text";
 import Ingredients from "../../elements/molecules/Ingredients";
 
 const VideoHeader = () => {
   const { class_id, role } = useParams();
   const [recipe, setRecipe] = useState();
 
-  const getData = useCallback(async () => {
-    const resp = apis.get_class_recipe({ class_id });
+  const getData = async () => {
+    const resp = await apis.get_class_recipe({ class_id });
+
     const { content } = resp.data;
     setRecipe(content);
-  }, [class_id]);
+  };
 
-  const leaveVideo = useCallback(async () => {
+  const leaveVideo = async () => {
     // publisher인 경우 종료 API 호출
     if (role === "pub") {
       const resp = await apis.quit_class({ class_id });
@@ -30,18 +30,18 @@ const VideoHeader = () => {
         return;
       }
     }
-  }, [class_id, role]);
+  };
 
   useEffect(() => {
     getData();
-    return () => leaveVideo();
-  }, [getData, leaveVideo]);
+    // return () => leaveVideo();
+  }, []);
 
   return (
     <>
       <StHeader>
-        <ST2>클래스 이름을 찾아 적어줍니다 어쩌고 저쩌고</ST2>
-        <div>상세 레시피 확인</div>
+        <StTitle>클래스 이름을 찾아 적어줍니다 어쩌고 저쩌고</StTitle>
+        <StRecipeButton>상세 레시피 확인</StRecipeButton>
       </StHeader>
       <StInfoPart>
         {recipe ? (
@@ -65,6 +65,26 @@ export default VideoHeader;
 
 const StHeader = styled.div`
   padding: 18px 18px 0px 18px;
+`;
+
+const StTitle = styled.div`
+  font-family: "Happiness Sans";
+  font-style: normal;
+  font-weight: 900;
+  font-size: 24px;
+  line-height: 30px;
+  color: #5b5b5b;
+`;
+
+const StRecipeButton = styled.button`
+  border: none;
+  font-family: "Noto Sans KR";
+  font-style: normal;
+  font-weight: 500;
+  font-size: 12px;
+  line-height: 16px;
+  text-decoration-line: underline;
+  color: #a5a5a5;
 `;
 
 const StInfoPart = styled.div``;
